requirements.txt
# FastAPI
fastapi==0.115.11

# Uvicorn ASGI 服务器
uvicorn==0.34.0

# 网页爬虫
playwright==1.40.0

# supabase
supabase==2.15.0

# python-dotenv
python-dotenv==1.1.0